useEffectの問題ファイルリスト取得関数を共通化
